+++ PagoElectronico/Clases/Transferencia.cs
@@ -103,8 +103,8 @@ namespace Clases
         public override void DataRowToObject(DataRow dr)
         {
             //this.transferencia_id = Convert.ToInt32(dr[""]);
-            this.CuentaOrigen.cuenta_id = Convert.ToDouble(dr["transferencia_origen_cuenta_id"]);
-            this.CuentaDestino.cuenta_id = Convert.ToDouble(dr["transferencia_destino_cuenta_id"]);
+            this.CuentaOrigen.cuenta_id = Convert.ToInt64(dr["transferencia_origen_cuenta_id"]);
+            this.CuentaDestino.cuenta_id = Convert.ToInt64(dr["transferencia_destino_cuenta_id"]);
             this.Importe = Convert.ToInt32(dr["transferencia_importe"]);
             this.Costo = Convert.ToInt32(dr["transferencia_costo"]);
             this.Fecha = Convert.ToDateTime(dr["transferencia_fecha"]);
